fixing test - test 1 passed
